iterator->streaming, split into iterable/iterator implementation, split infinite looping into it's own streamer
