--- docs/reference/catalog_iterator.rst
@@ -1,9 +1,0 @@
-Catalog Iteration
-=================
-
-.. currentmodule:: lsdb
-
-.. autosummary::
-    :toctree: api/
-
-    CatalogIterator
