+++ docs/reference/catalog_streaming.rst
@@ -1,0 +1,10 @@
+Catalog Streaming
+=================
+
+.. currentmodule:: lsdb.streams
+
+.. autosummary::
+    :toctree: api/
+
+    CatalogStream
+    InfiniteStream
