+++ src/lsdb/__init__.py
@@ -5,8 +5,8 @@ from .catalog import Catalog, MarginCatalog
 from .core.crossmatch.crossmatch import crossmatch
 from .core.search.region_search import BoxSearch, ConeSearch, PixelSearch, PolygonSearch
 from .io.show_versions import show_versions
-from .iterators.catalog_iterator import CatalogIterator
 from .loaders.dataframe.from_astropy import from_astropy
 from .loaders.dataframe.from_dataframe import from_dataframe
 from .loaders.hats.read_hats import open_catalog, read_hats
 from .nested.datasets import generate_catalog, generate_data
+from .streams.catalog_streams import CatalogStream, InfiniteStream
