+++ src/lsdb/streams/__init__.py
@@ -1,0 +1,1 @@
+from .catalog_streams import CatalogStream, InfiniteStream
