+++ tests/lsdb/streams/test_streams.py
@@ -1,0 +1,58 @@
+import pytest
+
+import lsdb
+
+
+def test_catalog_stream():
+    cat = lsdb.generate_catalog(100, 2, lowest_order=4, ra_range=(15.0, 25.0), dec_range=(34.0, 44.0), seed=1)
+
+    # Test default iteration
+    cat_stream = lsdb.CatalogStream(catalog=cat)
+
+    cat_iter = iter(cat_stream)
+    assert len(cat_iter.partitions_left) == cat.npartitions - cat_stream.partitions_per_chunk
+    assert len(cat_iter) == 11
+
+    total_len = 0
+    for chunk in cat_stream:
+        total_len += len(chunk)
+    assert total_len == len(cat)
+
+    # Test chunk size>1
+    cat_stream = lsdb.CatalogStream(catalog=cat, seed=1, partitions_per_chunk=2)
+    cat_iter = iter(cat_stream)
+    assert len(cat_iter.partitions_left) == cat.npartitions - cat_stream.partitions_per_chunk
+
+    total_len = 0
+    for chunk in cat_stream:
+        total_len += len(chunk)
+    assert total_len == len(cat)
+
+    # Test shuffing=False
+    cat_stream = lsdb.CatalogStream(catalog=cat, seed=1, shuffle_partitions=False)
+    cat_iter = iter(cat_stream)
+    assert len(cat_iter.partitions_left) == cat.npartitions - cat_stream.partitions_per_chunk
+    assert len(cat_iter) == 11
+
+    total_len = 0
+    for chunk in cat_stream:
+        total_len += len(chunk)
+    assert total_len == len(cat)
+
+
+def test_infinite_stream():
+    cat = lsdb.generate_catalog(100, 2, lowest_order=4, ra_range=(15.0, 25.0), dec_range=(34.0, 44.0), seed=1)
+    # Test infinite looping
+    cat_stream = lsdb.InfiniteStream(catalog=cat, seed=1)
+    cat_iter = iter(cat_stream)
+
+    # Check that we can sample beyond the number of partitions without error
+    for _ in range(cat.npartitions * 2):
+        next(cat_iter)
+
+
+def test_invalid_catalog_input():
+    with pytest.raises(
+        ValueError, match="The provided catalog input type <class 'str'> is not a lsdb.Catalog object."
+    ):
+        lsdb.CatalogStream(catalog="not a catalog")
